% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scraping.R
\name{clean_dataframe}
\alias{clean_dataframe}
\title{Clean the dataframe: keep only apartments, remove euro signs, extract the street..}
\usage{
clean_dataframe(castorus_data)
}
\arguments{
\item{castorus_data}{dataframe}
}
\value{
dataframe
}
\description{
Clean the dataframe: keep only apartments, remove euro signs, extract the street..
}
\examples{
url1 <- "https://www.castorus.com/s/Paris,51185,2-1----10000"
url2 <- "https://www.castorus.com/s/Paris,51185,2-1---10001-20000"
list_url <- list(url1, url2)
castorus_data <- get_dataframe(list_url)
clean_dataframe(castorus_data)
}
