% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_index.R
\name{column_index}
\alias{column_index}
\title{This function is used to be able to obtain the indexes of the selected rows in the datatable.}
\usage{
column_index(c, df1)
}
\arguments{
\item{c}{the indexes of the selected rows}

\item{df1}{the dataframe which is going to be used.}
}
\value{
data frame.
}
\description{
This function is used to be able to obtain the indexes of the selected rows in the datatable.
}
