% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_extraction.R
\name{combine_roads}
\alias{combine_roads}
\title{Combines the results form the datascrapping and the extraction from the descriptions}
\usage{
combine_roads(castorus, roadnames, stations)
}
\arguments{
\item{castorus}{dataframe}

\item{roadnames}{dataframe with street names}

\item{stations}{dataframe with station names}
}
\value{
dataframe with the apartments and their locations coming form the scraping and the extraction from the descriptions
}
\description{
Combines the results form the datascrapping and the extraction from the descriptions
}
