% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_slider.R
\name{filter_slider}
\alias{filter_slider}
\title{The Function inputs a dataframe and filters out the quantiles for a specific category that the user wants or do not want.}
\usage{
filter_slider(castorus_table, tri, min, max, value)
}
\arguments{
\item{castorus_table}{a data frame we are going to modify}

\item{tri}{The dataframe with the scores}

\item{min}{The miminum quantile we want.}

\item{max}{The highest quantile we want.}

\item{value}{Select the column in tri which is going to be filtered.}
}
\value{
data frame
}
\description{
The Function inputs a dataframe and filters out the quantiles for a specific category that the user wants or do not want.
}
