% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coordinate.R
\name{get_coordinate}
\alias{get_coordinate}
\title{Add coordinates of the apartment to the dataframe containing the ads information}
\usage{
get_coordinate(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe
}
\description{
Add coordinates of the apartment to the dataframe containing the ads information
}
