% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scraping.R
\name{get_dataframe}
\alias{get_dataframe}
\title{Create a dataframe with all the ads on Castorus website and the urls of their descriptions}
\usage{
get_dataframe(list_url)
}
\arguments{
\item{list_url}{list of urls}
}
\value{
dataframe
}
\description{
Create a dataframe with all the ads on Castorus website and the urls of their descriptions
}
\examples{
url1 <- "https://www.castorus.com/s/Paris,51185,2-1----10000"
url2 <- "https://www.castorus.com/s/Paris,51185,2-1---10001-20000"
list_url <- list(url1, url2)
get_dataframe(list_url)
}
