% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scraping.R
\name{get_description}
\alias{get_description}
\title{Get the description of an apartment}
\usage{
get_description(url)
}
\arguments{
\item{url}{url of a web page}
}
\value{
dataframe
}
\description{
Get the description of an apartment
}
\examples{
url <- "https://www.castorus.com/paris-5eme-ardsmnt,d63055365"
get_description(url)
}
