% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scraping.R
\name{get_link}
\alias{get_link}
\title{Create a dataframe with all the urls with the description page of the ads}
\usage{
get_link(url)
}
\arguments{
\item{url}{url of a web page}
}
\value{
dataframe
}
\description{
Create a dataframe with all the urls with the description page of the ads
}
\examples{
url <- "https://www.castorus.com/s/Paris,51185,2-1----10000"
get_link(url)
}
