% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_minimalfeatures.R
\name{get_minimalfeatures}
\alias{get_minimalfeatures}
\title{Extract the coordinates of the apartments with characteristics asked from the client}
\usage{
get_minimalfeatures(price, surface, rooms, N)
}
\arguments{
\item{price}{range of price of the apartment}

\item{surface}{range of surface of the apartment}

\item{rooms}{number of rooms in the apartment}

\item{N}{arrondissement.s in Paris}
}
\value{
dataframe with two columns longitude and latitude
}
\description{
Extract the coordinates of the apartments with characteristics asked from the client
}
\examples{
price <- c(300000,600000)
surface <- c(20,60)
N <- c(10,11)
rooms <- c(1,4,3)
get_minimalfeatures(price, surface, rooms, N)
}
