% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rank.R
\name{get_rank}
\alias{get_rank}
\title{Rank the apartments according to their proximity to a place}
\usage{
get_rank(dataframe, meters)
}
\arguments{
\item{dataframe}{dataframe}

\item{meters}{proximity in meters to the apartment}
}
\value{
dataframe with two columns: the index of the apartment and its rank
}
\description{
Rank the apartments according to their proximity to a place
}
