% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rank_restaurants.R
\name{get_rank_restaurants}
\alias{get_rank_restaurants}
\title{Rank the apartments according to their proximity to restaurants and bars}
\usage{
get_rank_restaurants(dataframe, meters)
}
\arguments{
\item{dataframe}{restaurants dataframe}

\item{meters}{proximity in meters to the apartment}
}
\value{
dataframe with two columns: the index of the apartment and its rank
}
\description{
Rank the apartments according to their proximity to restaurants and bars
}
