% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scraping.R
\name{get_table}
\alias{get_table}
\title{Create a dataframe from a table on a HTML page}
\usage{
get_table(url)
}
\arguments{
\item{url}{url of a web page}
}
\value{
dataframe
}
\description{
Create a dataframe from a table on a HTML page
}
\examples{
url <- "https://www.castorus.com/s/Paris,51185,2-1----10000"
get_table(url)
}
