% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_apartments.R
\name{join_dataframe}
\alias{join_dataframe}
\title{Score the apartments according to the densities of restaurants,commerce,cinema,schools and stations}
\usage{
join_dataframe(commerce, restaurants, cinema, schools, stations)
}
\arguments{
\item{commerce}{dataframe}

\item{restaurants}{dataframe}

\item{cinema}{dataframe}

\item{schools}{dataframe}

\item{stations}{dataframe}
}
\value{
dataframe with 1 column for the index of the good, and columns scoring the apartments among categories
}
\description{
Score the apartments according to the densities of restaurants,commerce,cinema,schools and stations
}
