% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_apartment.R
\name{location_apartment}
\alias{location_apartment}
\title{Choose the right function to locate the apartment according to the available information}
\usage{
location_apartment(num, street, metro)
}
\arguments{
\item{num}{number of the street of the apartment}

\item{street}{name of the street of the apartment}

\item{metro}{name of the nearest metro station}
}
\value{
dataframe with two columns : latitude and longitude
}
\description{
Choose the right function to locate the apartment according to the available information
}
\examples{
location_apartment(num= "12", street ="rue de Courcelles", metro = "Saint-Philippe-du-Roule")
location_apartment(num= "12", street ="rue de Courcelles", metro = NA)
location_apartment(num= NA, street ="rue de Courcelles", metro = "Saint-Philippe-du-Roule")
location_apartment(num= NA, street ="rue de Courcelles", metro = NA)
location_apartment(num= NA, street = NA, metro = "Saint-Philippe-du-Roule")
}
