% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_complete_address.R
\name{location_comp}
\alias{location_comp}
\title{Find the longitude and latitude of an apartment according to its complete address}
\usage{
location_comp(num, street)
}
\arguments{
\item{num}{number of the street of the apartment}

\item{street}{name of the street of the apartment}
}
\value{
dataframe with two columns : latitude and longitude
}
\description{
Find the longitude and latitude of an apartment according to its complete address
}
\examples{
localisation_comp("12", "rue de Courcelles")
}
