% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_metro.R
\name{location_metro}
\alias{location_metro}
\title{Find the longitude and latitude of an apartment according to the nearest metro station}
\usage{
location_metro(metro)
}
\arguments{
\item{metro}{name of the nearest metro station}
}
\value{
dataframe with two columns : latitude and longitude
}
\description{
Find the longitude and latitude of an apartment according to the nearest metro station
}
\examples{
location_metro("Saint-Philippe-du-Roule")
}
