% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_metro_street.R
\name{location_metro_street}
\alias{location_metro_street}
\title{Find the longitude and latitude of an apartment according to the nearest metro station and street}
\usage{
location_metro_street(metro, street)
}
\arguments{
\item{metro}{name of the nearest metro station}

\item{street}{name of the street of the apartment}
}
\value{
dataframe with two columns : latitude and longitude
}
\description{
Find the longitude and latitude of an apartment according to the nearest metro station and street
}
\examples{
location_metro_street("Saint-Philippe-du-Roule", "rue de Courcelles")
}
