% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_street.R
\name{location_street}
\alias{location_street}
\title{Find the longitude and latitude of an apartment according to its street}
\usage{
location_street(street)
}
\arguments{
\item{street}{street of the apartment}
}
\value{
dataframe with two columns : latitude and longitude
}
\description{
Find the longitude and latitude of an apartment according to its street
}
\examples{
location_street("rue de Courcelles")
}
