% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_extraction.R
\name{match_environnement}
\alias{match_environnement}
\title{extracts the roadnames or public transportation stations from the apartment descriptions}
\usage{
match_environnement(places, castorus)
}
\arguments{
\item{places}{dataframe with the names of Paris' streets or stations}

\item{castorus}{dataframe}
}
\value{
dataframe with the appartments and their streets or stations
}
\description{
extracts the roadnames or public transportation stations from the apartment descriptions
}
