% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_extraction.R
\name{number}
\alias{number}
\title{Extracts the street numbers from the description of the apartment}
\usage{
number(castorus)
}
\arguments{
\item{castorus}{dataframe}
}
\value{
dataframe with the apartments and their street numbers
}
\description{
Extracts the street numbers from the description of the apartment
}
\examples{
number(castorus)
}
