% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_apartments.R
\name{score_apartment}
\alias{score_apartment}
\title{Score the apartments according to the densities of restaurants,commerce,cinema,schools and stations}
\usage{
score_apartment(commerces, restaurants, cinemas, schools, stations)
}
\arguments{
\item{commerces}{dataframe}

\item{restaurants}{dataframe}

\item{cinemas}{dataframe}

\item{schools}{dataframe}

\item{stations}{dataframe}
}
\value{
dataframe with 1 column for the index of the good, and columns scoring the apartments among categories
}
\description{
Score the apartments according to the densities of restaurants,commerce,cinema,schools and stations
}
