% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scraping.R
\name{unaccent}
\alias{unaccent}
\title{Remove accents from a text}
\usage{
unaccent(text)
}
\arguments{
\item{text}{character}
}
\value{
character
}
\description{
Remove accents from a text
}
\examples{
unaccent("hétérogéneité")
}
