% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_extraction.R
\name{clean}
\alias{clean}
\title{Cleans the offers descriptions by removing the unwanted characters and getting rid of the ponctuation}
\usage{
clean(descriptions)
}
\arguments{
\item{descriptions}{character}
}
\value{
character
}
\description{
Cleans the offers descriptions by removing the unwanted characters and getting rid of the ponctuation
}
\examples{
annonces<-castorus$Description
clean(annonces)
}
