% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_minimalfeatures.R
\name{get_minimalfeatures}
\alias{get_minimalfeatures}
\title{Extract the coordinates of the apartments with characteristics asked from the client}
\usage{
get_minimalfeatures(min_price, max_price, min_surface, max_surface, N,
  min_room, max_room, castorus_table)
}
\arguments{
\item{min_price}{minimum price of the apartment}

\item{max_price}{maximum price of the apartment}

\item{min_surface}{minimum surface of the apartment}

\item{max_surface}{maximum surface of the apartment}

\item{N}{arrondissement in Paris}

\item{min_room}{minimum number of rooms}

\item{max_room}{maximum number of rooms}

\item{castorus_table}{dataframe with the informations}
}
\value{
dataframe with two columns longitude and latitude
}
\description{
Extract the coordinates of the apartments with characteristics asked from the client
}
